improve pager logic & types
