+++ src/assets/global.css
@@ -10,27 +10,3 @@ body,
     font-size: 20px;
     margin: 0;
 }
-
-#unsupportedmsg {
-    position: fixed;
-    top: 0;
-    left: 0;
-    right: 0;
-    bottom: 0;
-    text-align: center;
-    z-index: 9999;
-    background: #fff;
-    font-size: 2.5em;
-}
-
-@media (min-width: 1000px) {
-    #unsupportedmsg {
-        display: none;
-    }
-}
-
-@media all and (-ms-high-contrast: none), (-ms-high-contrast: active) {
-    #unsupportedmsg {
-        display: block;
-    }
-}
