+++ src/index.html
@@ -14,10 +14,6 @@ Pavel Malyshev, Evgeny Blokhin
     <title>OPTIMADE aggregated search</title>
 </head>
 
-<body>
-
-    <!-- <div id="unsupportedmsg">Sorry, your device or web-browser is not supported.</div> -->
-
-</body>
+<body />
 
 </html>
